refactor: change store folder name to context
