+++ src/components/Sidebar/ComparisonDetails/ComparisonDetails.tsx
@@ -6,7 +6,7 @@ import { Box } from '@mui/material';
 
 import Collapsible from '@components/Collapsible';
 
-import { useComparison } from '@store/comparisonContext';
+import { useComparison } from '@context/comparisonContext';
 import useMap from '@hook/useMap';
 
 import * as Styles from './styles';
