+++ src/components/Sidebar/Minimizer/Minimizer.tsx
@@ -4,7 +4,7 @@ import { FaChevronLeft, FaChevronRight } from 'react-icons/fa';
 
 import { useTheme } from '@mui/material/styles';
 
-import { useSidebar } from '@store/sidebarContext';
+import { useSidebar } from 'src/context/sidebarContext';
 
 import * as Styles from './styles';
